+++ CHANGELOG.md
@@ -5,12 +5,14 @@ It is based on Tidepool Blip 1.27.
 ## 3.9.0-beta - 2026-03-12
 ### Added
 - YLP-3613 Memos on DBL settings in Devices view
+- YLP-3614 Hcps can acknowledge patient monitoring alerts
 - YLP-3617 As a user, I can see the devices changes history
 - YLP-3785 Be able to know the effective DBLG1 or DBLG2 settings for the displayed day in the daily view
 - YLP-3813 Add hyperglycemia alerts (display in the patient view and settings in care teams and patients profile)
 - YLP-3837 Display Dexcom G7 sensor information in device view
 
 ### Changed
+- YLP-2478 Care team settings menu
 - YLP-2823 Use , instead of . for languages other than English
 - YLP-3618 Meal without bolus management in Daily view
 - YLP-3634 Add tooltips for Rescue carbs and Manual bolus metrics